// src/generated/prisma-client/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const ArticleStatus = {
  DRAFT: 'DRAFT',
  PENDING: 'PENDING',
  APPROVED: 'APPROVED',
  PUBLISHED: 'PUBLISHED',
  REJECTED: 'REJECTED'
} as const

export type ArticleStatus = (typeof ArticleStatus)[keyof typeof ArticleStatus]


export const ArticleCategory = {
  MARKETS_MACRO: 'MARKETS_MACRO',
  OPERATORS: 'OPERATORS',
  CAPITAL_STRATEGY: 'CAPITAL_STRATEGY',
  FAST_TAKE: 'FAST_TAKE',
  DEEP_DIVE: 'DEEP_DIVE'
} as const

export type ArticleCategory = (typeof ArticleCategory)[keyof typeof ArticleCategory]


export const SubscriberStatus = {
  ACTIVE: 'ACTIVE',
  UNSUBSCRIBED: 'UNSUBSCRIBED'
} as const

export type SubscriberStatus = (typeof SubscriberStatus)[keyof typeof SubscriberStatus]